Refactor: Campfire 컴포넌트 디렉토리 형식으로 변경
